# packages/shared/api-client/src/generated/model/healthVo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * monorepo-starter API
 * 企业级 monorepo starter 的 NestJS + Fastify API
 * OpenAPI spec version: 0.1.0
 */
import type { HealthVoStatus } from './healthVoStatus';

export interface HealthVo {
  status: HealthVoStatus;
  /** @minLength 1 */
  service: string;
  /** @minLength 1 */
  version: string;
}