# packages/shared/api-client/src/generated/model/updateUserDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * monorepo-starter API
 * 企业级 monorepo starter 的 NestJS + Fastify API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateUserDtoRole } from './updateUserDtoRole';
import type { UpdateUserDtoStatus } from './updateUserDtoStatus';

export interface UpdateUserDto {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username?: string;
  /** @pattern ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$ */
  email?: string;
  role?: UpdateUserDtoRole;
  status?: UpdateUserDtoStatus;
}